+++ apps/menu-bar/src/windows/index.ts
@@ -11,7 +11,7 @@ export const WindowsNavigator = createWindowsNavigator({
       windowStyle: {
         mask: [WindowStyleMask.Titled, WindowStyleMask.Closable],
         titlebarAppearsTransparent: true,
-        height: 660,
+        height: 680,
         width: 500,
       },
     },